Don't use non-existent keyword field.
With the type of `status` defined as `keyword` in https://github.com/codingexplained/complete-guide-to-elasticsearch/blame/master/Aggregations/introduction-to-aggregations.md#L31, the field `status.keyword` is not defined and the queries yield empty buckets. 

The proposed changes use the keyword field as intended. Alternatively the mapping could be adjusted. The associated lecture is contains the same error.